+++ odoo/local-src/sf_date_range/__manifest__.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+# Part of senseFly.
+
+{
+    'name': 'SenseFly Range',
+    'version': '1.0',
+    'author': 'Telmo Santos',
+    'website': 'http://www.sensefly.com',
+    'category': 'SenseFly',
+    'depends': ['date_range'],
+    'data': [
+        'data/date_range_type.xml',
+        'views/date_range_view.xml'],
+    'demo': [],
+    'images': ['static/description/icon.png'],
+    'installable': True,
+    'auto_install': False,
+}